Correct rubocop offense
